Remove ContextInterface & rename forks and threads
