+++ examples/thread.php
@@ -2,7 +2,7 @@
 <?php
 require dirname(__DIR__).'/vendor/autoload.php';
 
-use Icicle\Concurrent\Threading\ThreadContext;
+use Icicle\Concurrent\Threading\Thread;
 use Icicle\Coroutine;
 use Icicle\Loop;
 
@@ -14,7 +14,7 @@ $timer = Loop\periodic(1, function () {
 
 Coroutine\create(function () {
     // Create a new child thread that does some blocking stuff.
-    $context = new ThreadContext(function () {
+    $context = Thread::spawn(function () {
         printf("\$this: %s\n", get_class($this));
 
         printf("Received the following from parent: %s\n", (yield $this->receive()));
@@ -35,9 +35,6 @@ Coroutine\create(function () {
 
         yield 42;
     });
-
-    // Run the thread and wait asynchronously for it to finish.
-    $context->start();
 
     yield $context->send('Start data');
