--- src/ContextInterface.php
@@ -1,37 +1,0 @@
-<?php
-namespace Icicle\Concurrent;
-
-/**
- * Interface for all types of execution contexts.
- */
-interface ContextInterface extends ChannelInterface
-{
-    /**
-     * Checks if the context is running.
-     *
-     * @return bool True if the context is running, otherwise false.
-     */
-    public function isRunning();
-
-    /**
-     * Starts the context execution.
-     */
-    public function start();
-
-    /**
-     * Immediately kills the context.
-     */
-    public function kill();
-
-    /**
-     * @coroutine
-     *
-     * Gets a promise that resolves when the context ends and joins with the
-     * parent context.
-     *
-     * @return \Generator
-     *
-     * @resolve mixed Resolved with the return or resolution value of the context once it has completed execution.
-     */
-    public function join();
-}
